refactor(common): make parsing methods common
